Update request method to remove email requirement
We only send `email` to the request function to add it to the properties,
so this PR just updates the properties before sending it to the request.
This allows us to send any arbitrary data to the request - for example,
sending multiple items to the request body is now possible.